Positional magic 🌪: Airstrike
The air school is already quite positional: targeted spells require bolt
bouncing to use with maximum effectiveness, the remaining spells require
different positional considerations to use well.

Airstrike is smite-targeted, which makes it more interesting than most
single-target spells. However, it is just ``smite but air''. This commit
scales airstrike damage with the number of adjacent empty cells to the
target. Specifically, when there are more than three empty cells
adjacent to the target, each empty cell contributes a linear amount to the
affine base damage and reduces the variance by addinng another roll to
average the damage over.

At 100 spellpower this results in the following damage statistics,
depending on the number of empty adjacent spells.

|     | 0-3  |  4   |  5   |  6   |  7   |  8  |
| --- | ---- | ---- | ---- | ---- | ---- | --- |
| avg | 15.5 | 16.3 | 17.2 | 18.1 | 19.1 | 20  |
| std | 4.61 | 3.27 | 2.68 | 2.32 | 2.08 | 1.9 |

[ TODO: Refactor mons-cast airstrike damage to be in the same place as
  player damage. Take another swipe at the spell description. ]