Positional magic 🔥: Nerf conjure flame (minmay, tealizard)
Smite targeted, renewable flame clouds are extremely strong.

This changes Conjure flame to create the cloud on the player. The player
initially creates a cloud of smoldering embers. These ignite the turn
after if nothing is in the cell.

Conjure flame is still very strong and permits many of its old uses.
This increases the time cost of those uses and prevents degenerate
things like refereshing a blocking flame indefinitely.

It also did not work well as a monster spell; the only monsters that
still cast it were player ghosts, so it's been removed from them and is
now player only.