Positional magic ⭐️: Simplify Searing Ray (minmay)
Searing ray is a rather complicated spell because of beam aliasing, and
yet Cj is a hints mode background. This changes searing ray to always
fire at max range at the player's selected target (even if the target
moves), or in the player's selected direction. If the target dies the
ray continues firing at the target's last position. If there was no
monster, the spell proceedes like old searing ray (but fires for the
full distance).

In addition to the targeting change, the three beams are given the same
damage and to-hit, chosen so the damage distribution matches that of old
Searing Ray, and are all penetrating.