Positional magic 🐍: Adjust Venom Mage start
With Poison Magic being revised to emphasize transmutations and alchemy,
and Venom Bolt removed, this moves VM's starting Conjurations skill to
Transmutations and adds Ignite Poison to the Young Poisoner's Handbook.
This is especially relevant since Sting has become
Poison/Transmutations.

The manual and background descriptions are terse enough to not require a
change, but it might be nice for flavor reasons to add some reference to
alchemy.