Positional magic 🔥: New spell Foxfire
Foxfire is a L1 Conjurations/Fire spell that replaces Flame Tongue.
Foxfire conjures two foxfires adjacent to the caster
which seek the nearest target and burn the
target, or dissapate if they fail to reach the target after making 8 moves.
Foxfires cannot be swapped with.

Each foxfire burn does half the damage of Flame Tongue, but does not
miss. This makes the spell better than FT when fighting in the open
(discounting the delay), but much worse in corridors.

Foxfires have see invisible; with their high move speed this allows them
to attack first, dissipating and allowing the monster to then move
towards the player.

Preferably monsters would voluntarily walk into foxfire if it meant
getting to attack the player and trigger the foxfire attack in the same
way that the player swap occurs. Unfortunately this would significantly
complicate the already complicated monster-monster swap codepath.

The need arises from the fact that foxfires aren't normal summons and
are supposed to be projectiles, but weren't hitting invisible monsters
allowing them to be used as blockers.

An alternative approach to the issue, if SInv foxfires are too strong,
is to give foxfires low HP and a retaliatory attack like Shock Serpents.
That might prove more elegant in the long run. (The player swap
exception must remain.)

[Thanks to tealizard, Doesnt, EN0N, and cwz testers for feedback in
 development of this spell.]