Positional magic ⭐️: Revise Spellforged Servitor
Servitor now takes the highest-level servitor spell available for player
users. This could already be accomplished by a player by removing most
of their damaging spells and was a good idea. However
the process of modifying one's spell list introduced a variety of
tedious incentives in its interaction with amnesia.