Positional Magic ⛰️  💀: Rework BVC targeting
This commit brings Borgnjor's Vile Clutch targeting in line with most
other earth magic spells, making it use a directed beam instead of a
smite-targeted 3x3 square. The BVC beam always penetrates and never
misses, to affect up to 6 squares.

This is a nerf to a spell that was already sitting at a good power
level. However another aim of the 0.25 magic changes is to reduce the
overall strength of the Necromancy school given how many strong spells
it has. BVC is routinely overrated and not actually in the overpowered
set, so we don't want to make the spell bad just because of perception.
The reasons given for rating BVC so highly include synergy with allies,
lowering monster evasion, damage per cast, smite targeting, and being in
such a strong school. Allies are vastly stronger than BVC and capable of
clearing entire floors quite on their own, and the spell doesn't improve
accuracy enough versus a majority of monsters to warrant spend a turn
using BVC instead of a better Earth magic damage spell. Its damage per
cast is good, but at reasonable spellpower levels and in the most
important use cases, it's a lot better to spend MP each turn on a spell
that kills monsters sooner.

Removing the smite targeting combined with changes to other spells that
reduce the power of the Necromancy school will help with player
perception without making the spell too weak. I do also want to take a
look at the spell's damage per turn with the highest spellpower levels.

Part of the perception problem the spell has comes from constriction
having a lot of small effects that are hard for the player to evaluate.
If Necromancy settles on a kind of "melee support" role in 0.25, it
might be good to simplify the spell's effects somewhat toward this aim.